Refactoring Transpyler, moving namespace functions
